navisense-ml/requirements-lite.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow>=10.2.0
pinecone==8.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
boto3==1.34.34
requests>=2.32.3
sentence-transformers==2.2.2
